HS-42 REFACT: refactoring
