+++ Makefile
@@ -12,11 +12,10 @@ READLINE = -lreadline -L /usr/local/Cellar/readline/8.1.2/lib -I /usr/local/Cell
 
 RM = rm -f
 SRC = ./src/main.c ./src/welcome/print_wallpaper.c \
-./src/utils/ft_split_util.c ./src/utils/ft_split.c ./src/utils/ft_error.c \
-./src/utils/env.c ./src/utils/init.c ./src/utils/is_same_string.c ./src/utils/get_token_head.c \
-./src/utils/signal.c ./src/utils/safe_malloc.c \
-./src/built_in/ft_exit.c ./src/built_in/ft_env.c ./src/built_in/ft_export.c\
-./src/parser/make_arr_to_list.c ./src/parser/set_process_list.c ./src/temp_tester/command_list_test.c
+	  ./src/utils/ft_split.c ./src/utils/ft_split_util.c ./src/utils/ft_split_quote_util.c ./src/utils/ft_error.c ./src/utils/safe_malloc.c \
+	  ./src/utils/env.c ./src/utils/init.c ./src/utils/is_same_string.c ./src/utils/get_token_head.c ./src/utils/signal.c \
+	  ./src/built_in/ft_exit.c ./src/built_in/ft_env.c ./src/built_in/ft_export.c \
+	  ./src/parser/make_arr_to_list.c ./src/parser/set_process_list.c ./src/temp_tester/command_list_test.c
 
 OBJ=$(SRC:.c=.o)
