+++ include/minishell.h
@@ -35,7 +35,6 @@ enum e_token_type
 
 typedef struct s_token
 {
-	// int				quotes;
 	int				type;
 	char			*value;
 	struct s_token	*prev;
@@ -103,6 +102,7 @@ void	*safe_malloc(size_t size);
 void	print_error_message(char *str);
 void	ft_error_exit(char *str);
 void	get_new_prompt(int sig);
+int		handle_quote(t_split *split, char *line);
 
 /* parser */
 void	set_command_list(t_pslist **ps_list, t_token *tk_lst);
